Moves to Zipkin 2.18 module integration
Zipkin 2.18 no longer uses spring.factories to discover modules. We now
use yaml instead of AutoConfiguration. This results in smaller
distribution and startup times.
