+++ module/storage-stackdriver/src/main/java/zipkin/module/storage/stackdriver/CredentialsDecoratingClient.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin.autoconfigure.storage.stackdriver;
+package zipkin.module.storage.stackdriver;
 
 import com.google.auth.Credentials;
 import com.google.auth.RequestMetadataCallback;
