+++ module/storage-stackdriver/src/main/java/zipkin/module/storage/stackdriver/ZipkinStackdriverStorageProperties.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin.autoconfigure.storage.stackdriver;
+package zipkin.module.storage.stackdriver;
 
 import java.io.Serializable;
 import org.springframework.boot.context.properties.ConfigurationProperties;
